# optimized save_pointcloud_ply by 16.84x (1684%)