--- a/kornia/utils/pointcloud_io.py
+++ b/kornia/utils/pointcloud_io.py
@@ -16,6 +16,7 @@
 #
 
 import os
+
 import torch