--- a/tests/utils/test_pointcloud_io.py
+++ b/tests/utils/test_pointcloud_io.py
@@ -1,8 +1,28 @@
+# LICENSE HEADER MANAGED BY add-license-header
+#
+# Copyright 2018 Kornia Team
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 import os
-import pytest
+
 import numpy as np
+import pytest
 import torch
+
 import kornia
+
 from testing.base import assert_close  # assuming this is your custom assert_close