also wrote tests for the same, using more efficient numpy functions and removing unnecessary loops

--- Testing with 5000 points ---
Validation passed ✅
Original:  0.193949 sec
Optimized: 0.017146 sec
Speedup:   11.31×

--- Testing with 10000 points ---
Validation passed ✅
Original:  0.394249 sec
Optimized: 0.027022 sec
Speedup:   14.59×

--- Testing with 50000 points ---
Validation passed ✅
Original:  2.305624 sec
Optimized: 0.136922 sec
Speedup:   16.84×

https://colab.research.google.com/drive/1mYSBLYODzHbxlZtXH1Ygi92c5ibIb55j?usp=sharing

here's benchmarking and validadation code